+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## _(in-progress)_
+* New: Add ARIA attributes to improve accessibility
+    * Add `aria-controls` and `aria-expanded` attributes to excerpt toggle links
+    * Add `aria-hidden` attributes to comments excerpt and full `div`s as appropriate
 * Change: Add 'count' class to the "Expand/Collapse All" links to match styling of WP admin
 * Change: Add ELLIPSIS constant for ellipsis string and use it rather than harcoding string in various places
 * Change: Prevent translations from containing unintended markup